pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "asyncpg>=0.31.0",
    "deepgram-sdk>=6.0.1",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "numpy>=2.4.3",
    "openai>=2.28.0",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.9.0",
    "python-docx>=1.2.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "reportlab>=4.4.10",
    "sqlalchemy[asyncio]>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]
